requirements.txt
langchain
langchain-core
langchain-openai
langgraph
streamlit
python-dotenv
yfinance
